agent_py/requirements.txt
web3==6.15.1
python-dotenv==1.0.0
eth-abi==4.2.1
fastapi==0.115.0
uvicorn==0.32.1
